+++ unittest/std_pair.cpp
@@ -3,7 +3,6 @@
 
 #include <eigenpy/eigenpy.hpp>
 #include <eigenpy/std-pair.hpp>
-#include <iostream>
 
 namespace bp = boost::python;
 
@@ -17,6 +16,11 @@ std::pair<T1, T2> copy(const std::pair<T1, T2>& pair) {
   return pair;
 }
 
+template <typename T1, typename T2>
+const std::pair<T1, T2>& passthrough(const std::pair<T1, T2>& pair) {
+  return pair;
+}
+
 BOOST_PYTHON_MODULE(std_pair) {
   eigenpy::enableEigenPy();
 
@@ -25,4 +29,6 @@ BOOST_PYTHON_MODULE(std_pair) {
 
   bp::def("std_pair_to_tuple", std_pair_to_tuple<int, double>);
   bp::def("copy", copy<int, double>);
+  bp::def("passthrough", passthrough<int, double>,
+          bp::return_value_policy<bp::copy_const_reference>());
 }